pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
google-cloud-storage = "^3.0.0"
google-cloud-bigquery = "^3.29.0"
simplejson = "^3.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
